requirements.txt
django
django-environ
python-dotenv
djangorestframework
django-filter
django-cors-headers
drf-yasg
pillow

python-decouple 
requests

google-generativeai

psycopg2-binary

sendgrid
djangorestframework-simplejwt
stripe
